Module linkcheck::scanners

A scanner is just a function that which can extract links from a body of text.

Functions

markdown
A scanner that uses pulldown_cmark to extract all links from markdown.

markdown_with_broken_link_callback
A scanner that uses pulldown_cmark to extract all links from markdown, using the supplied callback to try and fix broken links.

plaintext
Use the linkify crate to find all URLs in a string of normal text.