Function linkcheck::scanners::markdown_with_broken_link_callback

pub fn markdown_with_broken_link_callback<'a>(
    src: &'a str, 
    on_broken_link: Option<&'a mut BrokenLinkCallback<'a>>
) -> impl Iterator<Item = (String, Span)> + 'a

A scanner that uses pulldown_cmark to extract all links from markdown, using the supplied callback to try and fix broken links.