Module linkcheck::validation

Code for validating the various types of Link.

Structs

BasicContext
A basic Context implementation which uses all the defaults.

Cache
A cache used to skip unnecessary network requests.

CacheEntry
A timestamped boolean used by the Cache to keep track of the last time a web crate::Link was checked.

InvalidLink
A Link and the Reason why it is invalid.

Options
Options to be used with resolve_link().

Outcomes
The result of validating a batch of Links.

Enums

Reason
Possible reasons for a bad link.

Traits

Context
Contextual information that callers can provide to guide the validation process.

Functions

check_filesystem
Check whether a Path points to a valid file on disk.

check_web
Check whether a Url points to a valid resource on the internet.

get	Deprecated
Send a HEAD request to a particular endpoint.

head
Send a HEAD request to a particular endpoint.

resolve_link
Try to resolve a link relative to the current directory.

validate
Validate several Links relative to a particular directory.