Struct linkcheck::validation::BasicContext

pub struct BasicContext {
    pub options: Options,
    // some fields omitted
}

A basic Context implementation which uses all the defaults.

Fields
options: Options

Options used when validating filesystem links.

Implementations

impl BasicContext

pub const USER_AGENT: &'static str

The User-Agent used by the BasicContext::client().

pub fn with_client(client: Client) -> Self

Create a BasicContext with an already initialized Client.

pub fn options_mut(&mut self) -> &mut Options

👎 Deprecated:
Access the field directly instead

Get a mutable reference to the Options used when validating filesystem links.